requirements.txt
fastapi==0.115.0
gunicorn==23.0.0
uvicorn==0.30.6
python-telegram-bot==20.9
groq==0.11.0
python-dotenv==1.0.1
slowapi==0.1.9
apscheduler==3.10.4
pytest==8.3.3
pytest-asyncio==0.24.0
